study-session/live-active-learning/requirements.txt
google-genai
google-cloud-storage
python-dotenv
fastapi
uvicorn
google-adk
websockets
psycopg2-binary